refactor rename histroy to history
