+++ src/main/java/qna/domain/question/Question.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 import javax.persistence.*;
 import java.util.List;
 import qna.domain.AbstractEntity;
-import qna.domain.histroy.ContentType;
-import qna.domain.histroy.DeleteHistory;
+import qna.domain.history.ContentType;
+import qna.domain.history.DeleteHistory;
 import qna.domain.user.User;
 
 @Entity
