+++ src/main/java/qna/service/DeleteHistoryService.java
@@ -3,8 +3,8 @@ package qna.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import qna.domain.histroy.DeleteHistory;
-import qna.domain.histroy.DeleteHistoryRepository;
+import qna.domain.history.DeleteHistory;
+import qna.domain.history.DeleteHistoryRepository;
 
 import javax.annotation.Resource;
 import java.util.List;
