+++ src/test/java/qna/domain/question/AnswerTest.java
@@ -4,12 +4,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDateTime;
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import qna.domain.UserTest;
-import qna.domain.histroy.ContentType;
-import qna.domain.histroy.DeleteHistory;
+import qna.domain.history.ContentType;
+import qna.domain.history.DeleteHistory;
 
 public class AnswerTest {
     public static final Answer JAVAJIGI = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
